requirements.txt
discord.py
matplotlib
pandas
python-dotenv
python-binance